pyproject.toml
[project]
name = "git-timemachine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "gitpython>=3.1.45",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.27",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
]
